requirements.txt
aiomqtt
aiohttp
bs4
lxml
imgkit
python-telegram-bot
python-telegram-bot[socks]